feat: configure request method + fetch image, collection Task

+++ URLSession_Example/URLSession_Example/Networks/EndPoints/CollectionEndPoint.swift
@@ -13,7 +13,7 @@ enum CollectionEndPoint {
 }
 
 extension CollectionEndPoint: EndPointable {
-    var requestTimeOut: Float {
+    var requestTimeOut: Double {
         return 20
     }
